backend/requirements.txt
-i https://pypi.org/simple
anyio==3.6.2 ; python_full_version >= '3.6.2'
asgiref==3.6.0 ; python_version >= '3.7'
click==8.1.3 ; python_version >= '3.7'
django==4.1.7
django-phonenumber-field[phonenumberslite]==7.0.2
ecdsa==0.18.0 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
fastapi==0.95.0
h11==0.14.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
passlib==1.7.4
phonenumbers==8.13.7
phonenumberslite==8.13.7
pillow==9.4.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pydantic==1.10.7 ; python_version >= '3.7'
pyjwt==2.6.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
sqlparse==0.4.3 ; python_version >= '3.5'
starlette==0.26.1 ; python_version >= '3.7'
typing-extensions==4.5.0 ; python_version >= '3.7'
uvicorn==0.21.1
